feat: separate 'update'
